Expose a method to update edit buttons without reinit
Containers which have had edit links added are now flagged with
dataset.wioAdded.

On setup(), all existing links are cleared, and then auth is
performed. If auth succeeds, links are added where appropriate, and
the containers are flagged. Meanwhile, auth information is cached. If
auth fails, cached auth information is cleared.

This means that now if auth disappears and setup() is run again, links
will be removed.

A new method update() is added to editbtn.js, which is exported from
index.js as updateEditButtons.

When the new method updateEditButtons() is called, nothing happens if
not already authed. But if authed, links are added to any new
appropriate DOM nodes. Nodes already flagged are skipped.